refactor: index.ts로 변경
indexRouter.ts 에서 index.ts 로 파일이름 변경
